+++ server/src/main/java/dev/food/fast/server/general/dto/request/ProductReviewRequest.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ProductReviewRequest {
     @JsonProperty("product_id")
-    private Integer idProduct;
+    private Long idProduct;
     private String review;
 }
