+++ server/src/main/java/dev/food/fast/server/general/repository/AddressOrderRepository.java
@@ -3,5 +3,5 @@ package dev.food.fast.server.general.repository;
 import dev.food.fast.server.general.models.order.AddressOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressOrderRepository extends JpaRepository<AddressOrder, Integer> {
+public interface AddressOrderRepository extends JpaRepository<AddressOrder, Long> {
 }
