+++ server/src/main/java/dev/food/fast/server/general/repository/BasicOrderRepository.java
@@ -3,5 +3,5 @@ package dev.food.fast.server.general.repository;
 import dev.food.fast.server.general.models.order.BasicOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BasicOrderRepository extends JpaRepository<BasicOrder, Integer> {
+public interface BasicOrderRepository extends JpaRepository<BasicOrder, Long> {
 }
