+++ server/src/main/java/dev/food/fast/server/general/repository/BasicOrderUserRepository.java
@@ -3,5 +3,5 @@ package dev.food.fast.server.general.repository;
 import dev.food.fast.server.general.models.order.BasicOrderUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BasicOrderUserRepository extends JpaRepository<BasicOrderUser, Integer> {
+public interface BasicOrderUserRepository extends JpaRepository<BasicOrderUser, Long> {
 }
