+++ server/src/main/java/dev/food/fast/server/general/repository/ProductsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
-    Optional<Product> findById(Integer id);
+public interface ProductsRepository extends JpaRepository<Product, Long> {
+    Optional<Product> findById(Long id);
 }
