+++ server/src/main/java/dev/food/fast/server/general/service/ProductReviewService.java
@@ -68,7 +68,7 @@ public class ProductReviewService {
         }
     }
 
-    public ResponseEntity<?> getAllProductReview(Integer productId) {
+    public ResponseEntity<?> getAllProductReview(Long productId) {
         var productLikesResponseList = productReviewRepository.findProductReviewByProduct_Id(productId);
 
         if (productLikesResponseList.isEmpty()) {
